*BDS port:
- split *BDS files
- add CMake confingure instructions on bsd systems
- add includes to main file
