+++ CMakeLists.txt
@@ -34,11 +34,13 @@ ELSEIF(CMAKE_SYSTEM_NAME MATCHES "Darwin")
 ELSEIF(CMAKE_SYSTEM_NAME MATCHES "FreeBSD")
    # FreeBSD STUFF HERE
    message( "FreeBSD detected")
-   message( FATAL_ERROR "Free BSD is not supported yet" )
+   message( WARNING "FreeBSD is still experimental!" )
+   SET( METER_SOURCES "bsd/memory_freebsd.cc" "bsd/cpu.cc" "bsd/load.cc" )
 ELSEIF(CMAKE_SYSTEM_NAME MATCHES "OpenBSD")
    # OpenBSD Stuff Here
    message( "OpenBSD detected")
-   message( FATAL_ERROR "OpenBSD is not supported yet" )
+   message( WARNING "OpenBSD is still experimental!" )
+   SET( METER_SOURCES "bsd/memory_openbsd.cc" "bsd/cpu.cc" "bsd/load.cc" )
 ELSE()
    message( FATAL_ERROR "Cannot be compiled on this system" )
 endif()
