+++ bsd/cpu.h
@@ -1,0 +1,6 @@
+#ifndef CPU_H_
+#define CPU_H_
+
+float cpu_percentage(unsigned);
+
+#endif
