+++ bsd/load.h
@@ -1,0 +1,8 @@
+#ifndef LOAD_H_
+#define LOAD_H_
+
+#include <string>
+
+std::string load_string( bool );
+
+#endif
