+++ bsd/memory.h
@@ -1,0 +1,8 @@
+#ifndef MEMORY_H_
+#define MEMORY_H_
+
+#include <string>
+
+std::string mem_string( bool );
+
+#endif
